+++ example/testcontent/models.py
@@ -1,11 +1,14 @@
 """Models for testing an example Django App."""
 from django.db import models
 
+from elasticsearch_dsl import field
+
 from djbetty.fields import ImageField
 
 from bulbs.campaigns.models import Campaign
 from bulbs.content.models import Content, Tag, ElasticsearchImageField
 from bulbs.reading_list.mixins import ReadingListMixin
+from bulbs.recirc.mixins import BaseQueryMixin
 
 
 class TestContentObj(Content):
@@ -79,3 +82,12 @@ class TestContentDetailImage(TestContentObj):
     def get_serializer_class(cls):
         from .serializers import TestContentDetailImageSerializer
         return TestContentDetailImageSerializer
+
+
+class TestRecircContentObject(Content, BaseQueryMixin):
+
+    foo = models.CharField(max_length=255)
+    bar = models.CharField(max_length=255)
+
+    class Mapping(Content.Mapping):
+        query = field.Object(enabled=False)