+++ tests/content/test_content_views.py
@@ -1,3 +1,4 @@
+import time
 from datetime import timedelta
 
 from django.contrib.auth import get_user_model
@@ -60,13 +61,16 @@ class TestContentViews(BaseIndexableTestCase):
         # create test content and token
         create_date = timezone.now()
         expire_date = create_date + timedelta(days=3)
-        content = make_content(published=None)
+        content = make_content(TestContentObj, published=None)
         obfuscated_url_info = ObfuscatedUrlInfo.objects.create(
             content=content,
             create_date=create_date.isoformat(),
             expire_date=expire_date.isoformat()
         )
         uuid = obfuscated_url_info.url_uuid
+
+        time.sleep(1)
+        Content.search_objects.refresh()
 
         # attempt to get article via token
         response = self.client.get(reverse("unpublished", kwargs={"token": uuid}))